# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Authenticate user",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Revoke user authentication",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/points": {
            "get": {
                "tags": [
                    "Points"
                ],
                "description": "Display a list of records.",
                "operationId": "1cf460f6763a64fef3faa484f8400f58",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search for name/register/cpf/id/user_id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Points"
                ],
                "description": "Store a record in the database.",
                "operationId": "26ded31f2bd793baa7548aab9e886056",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "register": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 37.7749
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": -122.4194
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "base64"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/points/{id}": {
            "put": {
                "tags": [
                    "Points"
                ],
                "description": "Update the specified record in the database.",
                "operationId": "307cb07784f48ef7fb1f01c1e5424944",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Point ID to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "register": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 37.7749
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": -122.4194
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "base64"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Points"
                ],
                "description": "Delete a user based on the given ID",
                "operationId": "5510e2920d2c28770c5a9109ae1d39f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Point ID to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}